refactor: Use "for await" to read stdin from the client
It seems that the runtime side stdio is unstable.
- The code to close stdin (Writable) is not implemented
- stdin.write Promise does not complete

+++ src/extension/lib/handleRun.ts
@@ -97,15 +97,19 @@ export class HandleRun implements IpcHandler {
           options
         )
       }
-      request.pipeIn?.on('data', (data) => {
-        pipeIn.write(data)
-      })
-      request.pipeIn?.once('close', async () => {
+      ;(async () => {
+        for await (const data of request.pipeIn!) {
+          // await pipeIn.write(data) // await が返ってこない.
+          pipeIn.write(data)
+          if (process === undefined) break
+        }
         // https://github.com/microsoft/vscode-wasm/issues/143
-        // この辺の暫定的な対応。
-        await new Promise((resolve) => setTimeout(resolve, 1000 * 5))
-        process?.terminate()
-      })
+        // この辺の暫定的な対応にしたかったが stdio 周りはまだ安定していないようなのであきらめる
+        if (process !== undefined) {
+          await new Promise((resolve) => setTimeout(resolve, 1000 * 5))
+          process?.terminate()
+        }
+      })()
       const started = Date.now()
       exitStatus = await process.run()
       if (request.args.runArgs['print-elapsed-time']) {
